Factor out ChunkDelayQueue into own file
